Allow user to input desired width of grid
Added a text box and a submit button for the user to type in a number
and click submit. After clicking submit, the original grid is deleted
and replaced by a new grid with the width/height the user inputted.

Existing code was placed into functions: gridSetup(), sketching(), and
deleteGrid().